Move tests to new place
